Changes to meshes

+++ LUtils/Geometry/Mesh.h
@@ -6,7 +6,7 @@ namespace LUtils{
 	namespace Geometry {
 		class Mesh {
 		public:
-			Mesh(std::vector<Vertex> _vertices, std::vector<unsigned int> _indices) {
+			Mesh(std::vector<Vertex> _vertices, std::vector<size_t> _indices) {
 				this->mVertices = _vertices;
 				this->mIndices = _indices;
 			};
@@ -14,11 +14,11 @@ namespace LUtils{
 			~Mesh() {};
 
 			std::vector<Vertex> GetVertices() { return this->mVertices; };
-			std::vector<unsigned int> GetIndices() { return this->mIndices; };
+			std::vector<size_t> GetIndices() { return this->mIndices; };
 
 		private:
 			std::vector<Vertex> mVertices;
-			std::vector<unsigned int> mIndices;
+			std::vector<size_t> mIndices;
 		};
 	}
 }
